requirements.txt
fastapi
uvicorn
psycopg
pytest
httpx